# apps/web/src/content/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ClearGateway API",
    "version": "1.0.0",
    "description": "Non-custodial stablecoin payments (testnet)."
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      },
      "adminAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Internal API: ADMIN_TOKEN, plus an X-Admin-Email header naming a staff account."
      }
    },
...
  },
  "paths": {
...
    "/v1/refunds": {
      "post": {
        "summary": "Record a refund request",
        "tags": [
          "refunds"
        ],
        "description": "Records the request only: nothing is sent on-chain yet. `amount` (USDC base units) defaults to everything still refundable. The total of all live refunds can never exceed the amount the customer paid. Emits `refund.created`.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "payment_intent_id": {
                    "type": "string",
                    "pattern": "^pi_[0-9a-f]{32}$"
                  },
                  "amount": {
                    "type": "string",
                    "pattern": "^[1-9]\\d{0,17}$"
                  },
                  "to_address": {
                    "type": "string"
                  },
                  "reason": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 500
                  }
                },
                "required": [
                  "payment_intent_id",
                  "to_address"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 255
            },
            "in": "header",
            "name": "idempotency-key",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "object": {
                      "type": "string",
                      "enum": [
                        "refund"
                      ]
                    },
                    "payment_intent_id": {
                      "type": "string"
                    },
                    "amount": {
                      "type": "string"
                    },
                    "currency": {
                      "type": "string"
                    },
                    "decimals": {
                      "type": "number"
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "requested",
                        "approved",
                        "processing",
                        "succeeded",
                        "failed",
                        "rejected"
                      ]
                    },
                    "to_address": {
                      "type": "string"
                    },
                    "tx_hash": {
                      "type": "string",
                      "nullable": true
                    },
                    "reason": {
                      "type": "string",
                      "nullable": true
                    },
                    "requested_by": {
                      "type": "string"
                    },
                    "created_at": {
                      "type": "string"
                    },
                    "updated_at": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "object",
                    "payment_intent_id",
                    "amount",
                    "currency",
                    "decimals",
                    "status",
                    "to_address",
                    "tx_hash",
                    "reason",
                    "requested_by",
                    "created_at",
                    "updated_at"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List your refunds (newest first, cursor pagination)",
        "tags": [
          "refunds"
        ],
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            },
            "in": "query",
            "name": "limit",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100
            },
            "in": "query",
            "name": "starting_after",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "pattern": "^pi_[0-9a-f]{32}$"
            },
            "in": "query",
            "name": "payment_intent_id",
            "required": false
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "object": {
                      "type": "string",
                      "enum": [
                        "list"
                      ]
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "object": {
                            "type": "string",
                            "enum": [
                              "refund"
                            ]
                          },
                          "payment_intent_id": {
                            "type": "string"
                          },
                          "amount": {
                            "type": "string"
                          },
                          "currency": {
                            "type": "string"
                          },
                          "decimals": {
                            "type": "number"
                          },
                          "status": {
                            "type": "string",
                            "enum": [
                              "requested",
                              "approved",
                              "processing",
                              "succeeded",
                              "failed",
                              "rejected"
                            ]
                          },
                          "to_address": {
                            "type": "string"
                          },
                          "tx_hash": {
                            "type": "string",
                            "nullable": true
                          },
                          "reason": {
                            "type": "string",
                            "nullable": true
                          },
                          "requested_by": {
                            "type": "string"
                          },
                          "created_at": {
                            "type": "string"
                          },
                          "updated_at": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "object",
                          "payment_intent_id",
                          "amount",
                          "currency",
                          "decimals",
                          "status",
                          "to_address",
                          "tx_hash",
                          "reason",
                          "requested_by",
                          "created_at",
                          "updated_at"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "has_more": {
                      "type": "boolean"
                    },
                    "next_cursor": {
                      "type": "string",
                      "nullable": true
                    }
                  },
                  "required": [
                    "object",
                    "data",
                    "has_more",
                    "next_cursor"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  }
}